ai/requirements.txt
monai[all]==1.3.0
highdicom==0.22.0
pydicom==2.4.3
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
Pillow>=9.5.0
requests>=2.31.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
